+++ dwds/lib/src/debugging/inspector.dart
@@ -4,8 +4,9 @@
 
 import 'dart:math' as math show min;
 
+import 'package:dwds/src/debugging/remote_debugger.dart';
 import 'package:path/path.dart' as p;
-import 'package:vm_service_lib/vm_service_lib.dart';
+import 'package:vm_service/vm_service.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 import '../services/chrome_proxy_service.dart';
@@ -38,7 +39,7 @@ class AppInspector extends Domain {
   /// Map of [ScriptRef] id to containing [LibraryRef] id.
   final _scriptIdToLibraryId = <String, String>{};
 
-  final WipDebugger _wipDebugger;
+  final RemoteDebugger _remoteDebugger;
   final AssetHandler _assetHandler;
   final Debugger debugger;
   final Isolate isolate;
@@ -52,7 +53,7 @@ class AppInspector extends Domain {
     this._assetHandler,
     this.debugger,
     this._root,
-    this._wipDebugger,
+    this._remoteDebugger,
   )   : isolateRef = _toIsolateRef(isolate),
         super.forInspector();
 
@@ -81,7 +82,7 @@ class AppInspector extends Domain {
     ..number = isolate.number;
 
   static Future<AppInspector> initialize(
-    WipDebugger wipDebugger,
+    RemoteDebugger remoteDebugger,
     AssetHandler assetHandler,
     Debugger debugger,
     String root,
@@ -96,7 +97,7 @@ class AppInspector extends Domain {
       ..libraries = []
       ..extensionRPCs = [];
     var inspector =
-        AppInspector._(isolate, assetHandler, debugger, root, wipDebugger);
+        AppInspector._(isolate, assetHandler, debugger, root, remoteDebugger);
     await inspector._initialize();
     return inspector;
   }
@@ -160,7 +161,7 @@ class AppInspector extends Domain {
   Future<RemoteObject> _callFunctionOn(
       RemoteObject receiver, String evalExpression, List arguments) async {
     var result =
-        await _wipDebugger.sendCommand('Runtime.callFunctionOn', params: {
+        await _remoteDebugger.sendCommand('Runtime.callFunctionOn', params: {
       'functionDeclaration': evalExpression,
       'arguments': arguments,
       'objectId': receiver.objectId,
@@ -233,7 +234,7 @@ class AppInspector extends Domain {
   Future<RemoteObject> evaluateJsExpression(String expression) async {
     // TODO(alanknight): Support a version with arguments if needed.
     WipResponse result;
-    result = await _wipDebugger
+    result = await _remoteDebugger
         .sendCommand('Runtime.evaluate', params: {'expression': expression});
     handleErrorIfPresent(result, evalContents: expression, additionalDetails: {
       'Dart expression': expression,
@@ -259,7 +260,7 @@ function($argsString) {
 }
     ''';
     var result =
-        await _wipDebugger.sendCommand('Runtime.callFunctionOn', params: {
+        await _remoteDebugger.sendCommand('Runtime.callFunctionOn', params: {
       'functionDeclaration': evalExpression,
       'arguments': arguments,
       // TODO(jakemac): Use the executionContext instead, or possibly the
@@ -384,7 +385,7 @@ function($argsString) {
       return result;
     })()
     ''';
-    var result = await _wipDebugger.sendCommand('Runtime.evaluate',
+    var result = await _remoteDebugger.sendCommand('Runtime.evaluate',
         params: {'expression': expression, 'returnByValue': true});
     handleErrorIfPresent(result, evalContents: expression);
     var classDescriptors = (result.result['result']['value']['classes'] as List)
@@ -524,7 +525,7 @@ function($argsString) {
   Future<List<LibraryRef>> _getLibraryRefs() async {
     if (_libraryRefs.isNotEmpty) return _libraryRefs.values.toList();
     var expression = "require('dart_sdk').dart.getLibraries();";
-    var librariesResult = await _wipDebugger.sendCommand('Runtime.evaluate',
+    var librariesResult = await _remoteDebugger.sendCommand('Runtime.evaluate',
         params: {'expression': expression, 'returnByValue': true});
     handleErrorIfPresent(librariesResult, evalContents: expression);
     var libraries =
@@ -542,7 +543,7 @@ function($argsString) {
   /// Runs an eval on the page to compute all existing registered extensions.
   Future<List<String>> _getExtensionRpcs() async {
     var expression = "require('dart_sdk').developer._extensions.keys.toList();";
-    var extensionsResult = await _wipDebugger.sendCommand('Runtime.evaluate',
+    var extensionsResult = await _remoteDebugger.sendCommand('Runtime.evaluate',
         params: {'expression': expression, 'returnByValue': true});
     handleErrorIfPresent(extensionsResult, evalContents: expression);
     return List.from(extensionsResult.result['result']['value'] as List);