+++ dwds/lib/src/handlers/dev_handler.dart
@@ -7,6 +7,9 @@ import 'dart:convert';
 
 import 'package:build_daemon/data/build_status.dart';
 import 'package:build_daemon/data/serializers.dart' as build_daemon;
+import 'package:dwds/src/debugging/remote_debugger.dart';
+import 'package:dwds/src/debugging/webkit_debugger.dart';
+import 'package:dwds/src/servers/extension_backend.dart';
 import 'package:logging/logging.dart';
 import 'package:pedantic/pedantic.dart';
 import 'package:shelf/shelf.dart';
@@ -39,6 +42,7 @@ class DevHandler {
   final bool _verbose;
   final void Function(Level, String) _logWriter;
   final Future<ChromeConnection> Function() _chromeConnection;
+  final ExtensionBackend _extensionBackend;
 
   Stream<AppConnection> get connectedApps => _connectedApps.stream;
 
@@ -50,9 +54,13 @@ class DevHandler {
     this._hostname,
     this._verbose,
     this._logWriter,
+    this._extensionBackend,
   ) {
     _sub = buildResults.listen(_emitBuildResults);
     _listen();
+    if (_extensionBackend != null) {
+      _startExtensionDebugService();
+    }
   }
 
   Handler get handler => _sseHandler.handler;
@@ -102,11 +110,11 @@ class DevHandler {
 
     await tabConnection.runtime.enable();
 
-    var wipDebugger = WipDebugger(tabConnection);
+    var webkitDebugger = WebkitDebugger(WipDebugger(tabConnection));
 
     return DebugService.start(
       _hostname,
-      wipDebugger,
+      webkitDebugger,
       appTab.url,
       _assetHandler.getRelativeAsset,
       appInstanceId,
@@ -125,8 +133,7 @@ class DevHandler {
         var debugService =
             await startDebugService(await _chromeConnection(), instanceId);
         var appServices = await _createAppDebugServices(appId, debugService);
-        unawaited(appServices
-            .chromeProxyService.wipDebugger.connection.onClose.first
+        unawaited(appServices.chromeProxyService.remoteDebugger.onClose.first
             .whenComplete(() {
           appServices.close();
           _servicesByAppId.remove(appId);
@@ -195,8 +202,8 @@ class DevHandler {
 
         // If you load the same app in a different tab then we need to throw
         // away our old services and start new ones.
-        if (!(await _isCorrectTab(
-            message.instanceId, appServices.chromeProxyService.wipDebugger))) {
+        if (!(await _isCorrectTab(message.instanceId,
+            appServices.chromeProxyService.remoteDebugger))) {
           unawaited(appServices.close());
           unawaited(_servicesByAppId.remove(message.appId));
           appServices =
@@ -207,7 +214,7 @@ class DevHandler {
             serializers.serialize(DevToolsResponse((b) => b..success = true))));
 
         appServices.connectedInstanceId = message.instanceId;
-        await appServices.chromeProxyService.wipDebugger
+        await appServices.chromeProxyService.remoteDebugger
             .sendCommand('Target.createTarget', params: {
           'newWindow': true,
           'url': 'http://${_devTools.hostname}:${_devTools.port}'
@@ -228,7 +235,7 @@ class DevHandler {
           // Re-connect to the previous instance if its in the same tab,
           // otherwise do nothing for now.
           if (await _isCorrectTab(
-              message.instanceId, services.chromeProxyService.wipDebugger)) {
+              message.instanceId, services.chromeProxyService.remoteDebugger)) {
             services.connectedInstanceId = message.instanceId;
             await services.chromeProxyService.createIsolate();
           }
@@ -272,11 +279,41 @@ class DevHandler {
     var webdevClient = await DwdsVmClient.create(debugService);
     return AppDebugServices(debugService, webdevClient);
   }
+
+  /// Starts a [DebugService] for Dart Debug Extension.
+  void _startExtensionDebugService() async {
+    var _extensionDebugger = await _extensionBackend.extensionDebugger;
+    // Waits for a `DevToolsRequest` to be sent from the extension background
+    // when the extension is clicked.
+    // TODO(pisong): Handle multiple `devToolsRequest`.
+    await _extensionDebugger.devToolsRequestStream.first;
+    var debugService = await DebugService.start(
+      _hostname,
+      _extensionDebugger,
+      _extensionDebugger.tabUrl,
+      _assetHandler.getRelativeAsset,
+      _extensionDebugger.appId,
+      onResponse: _verbose
+          ? (response) {
+              if (response['error'] == null) return;
+              _logWriter(Level.WARNING,
+                  'VmService proxy responded with an error:\n$response');
+            }
+          : null,
+    );
+    var appServices =
+        await _createAppDebugServices(_extensionDebugger.appId, debugService);
+    await _extensionDebugger.sendCommand('Target.createTarget', params: {
+      'newWindow': true,
+      'url': 'http://${_devTools.hostname}:${_devTools.port}'
+          '/?hide=none&uri=${appServices.debugService.wsUri}',
+    });
+  }
 }
 
-/// Checks if connection of [wipDebugger] is running the app with [instanceId].
-Future<bool> _isCorrectTab(String instanceId, WipDebugger wipDebugger) async {
-  var result = await wipDebugger.connection.runtime
-      .evaluate(r'window["$dartAppInstanceId"];');
+/// Checks if connection of [remoteDebugger] is running the app with [instanceId].
+Future<bool> _isCorrectTab(
+    String instanceId, RemoteDebugger remoteDebugger) async {
+  var result = await remoteDebugger.evaluate(r'window["$dartAppInstanceId"];');
   return result.value == instanceId;
 }