+++ dwds/lib/src/servers/extension_debugger.dart
@@ -3,16 +3,19 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:collection';
 import 'dart:convert';
 
 import 'package:dwds/data/devtools_request.dart';
 import 'package:dwds/data/extension_request.dart';
 import 'package:dwds/data/serializers.dart';
+import 'package:dwds/src/debugging/remote_debugger.dart';
 import 'package:sse/server/sse_handler.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
-/// A debugger backed by the Dart Debug Extension.
-class ExtensionDebugger implements WipDebugger {
+/// A remote debugger backed by the Dart Debug Extension
+/// with an SSE connection.
+class ExtensionDebugger implements RemoteDebugger {
   /// A connection between the debugger and the background of
   /// Dart Debug Extension
   final SseConnection sseConnection;
@@ -21,9 +24,6 @@ class ExtensionDebugger implements WipDebugger {
   final _completers = <int, Completer>{};
   final _eventStreams = <String, Stream>{};
   var _completerId = 0;
-
-  @override
-  WipConnection get connection => throw UnimplementedError();
 
   String tabUrl;
   String appId;
@@ -36,12 +36,16 @@ class ExtensionDebugger implements WipDebugger {
   final _notificationController = StreamController<WipEvent>.broadcast();
   Stream<WipEvent> get onNotification => _notificationController.stream;
 
+  @override
   Stream<ConsoleAPIEvent> get onConsoleAPICalled => eventStream(
       'Runtime.consoleAPICalled', (WipEvent event) => ConsoleAPIEvent(event));
 
+  @override
   Stream<ExceptionThrownEvent> get onExceptionThrown => eventStream(
       'Runtime.exceptionThrown',
       (WipEvent event) => ExceptionThrownEvent(event));
+
+  final _scripts = <String, WipScript>{};
 
   ExtensionDebugger(this.sseConnection) {
     sseConnection.stream.listen((data) {
@@ -70,6 +74,13 @@ class ExtensionDebugger implements WipDebugger {
     }, onError: (_) {
       close();
     });
+    onScriptParsed.listen((event) {
+      _scripts[event.script.scriptId] = event.script;
+    });
+    // Listens for a page reload.
+    onGlobalObjectCleared.listen((_) {
+      _scripts.clear();
+    });
   }
 
   /// Sends a [command] with optional [params] to Dart Debug Extension
@@ -90,7 +101,12 @@ class ExtensionDebugger implements WipDebugger {
 
   int newId() => _completerId++;
 
-  Future<void> close() => sseConnection.sink.close();
+  @override
+  void close() {
+    sseConnection.sink.close();
+    _notificationController.close();
+    _devToolsRequestController.close();
+  }
 
   @override
   Future disable() => sendCommand('Debugger.disable');
@@ -125,6 +141,13 @@ class ExtensionDebugger implements WipDebugger {
   Future<WipResponse> stepOver() => sendCommand('Debugger.stepOver');
 
   @override
+  Future<void> enablePage() => throw UnimplementedError();
+
+  @override
+  Future<RemoteObject> evaluate(String expression) =>
+      throw UnimplementedError();
+
+  @override
   Stream<T> eventStream<T>(String method, WipEventTransformer<T> transformer) {
     return _eventStreams
         .putIfAbsent(
@@ -136,11 +159,8 @@ class ExtensionDebugger implements WipDebugger {
   }
 
   @override
-  Stream<WipDomain> get onClosed => throw UnimplementedError();
-
-  @override
   Stream<GlobalObjectClearedEvent> get onGlobalObjectCleared => eventStream(
-      'Debugger.globalObjectCleared',
+      'Page.frameStartedLoading',
       (WipEvent event) => GlobalObjectClearedEvent(event));
 
   @override
@@ -150,12 +170,16 @@ class ExtensionDebugger implements WipDebugger {
   @override
   Stream<DebuggerResumedEvent> get onResumed => eventStream(
       'Debugger.resumed', (WipEvent event) => DebuggerResumedEvent(event));
+
   @override
   Stream<ScriptParsedEvent> get onScriptParsed => eventStream(
       'Debugger.scriptParsed', (WipEvent event) => ScriptParsedEvent(event));
 
   @override
-  Map<String, WipScript> get scripts => throw UnimplementedError();
+  Map<String, WipScript> get scripts => UnmodifiableMapView(_scripts);
+
+  @override
+  Stream<WipConnection> get onClose => throw UnimplementedError();
 
   String _pauseStateToString(PauseState state) {
     switch (state) {