+++ dwds/test/chrome_proxy_service_test.dart
@@ -13,7 +13,7 @@ import 'package:dwds/src/utilities/dart_uri.dart';
 import 'package:http/http.dart' as http;
 import 'package:pedantic/pedantic.dart';
 import 'package:test/test.dart';
-import 'package:vm_service_lib/vm_service_lib.dart';
+import 'package:vm_service/vm_service.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
 import 'fixtures/context.dart';
@@ -170,20 +170,12 @@ void main() {
       });
     });
 
-    test('clearCpuProfile', () {
-      expect(() => service.clearCpuProfile(null), throwsUnimplementedError);
-    });
-
     test('clearVMTimeline', () {
       expect(() => service.clearVMTimeline(), throwsUnimplementedError);
     });
 
     test('clearVMTimeline', () {
       expect(() => service.clearVMTimeline(), throwsUnimplementedError);
-    });
-
-    test('collectAllGarbage', () {
-      expect(() => service.collectAllGarbage(null), throwsUnimplementedError);
     });
 
     test('clearVMTimeline', () {
@@ -294,10 +286,6 @@ void main() {
     test('getAllocationProfile', () {
       expect(
           () => service.getAllocationProfile(null), throwsUnimplementedError);
-    });
-
-    test('getCpuProfile', () {
-      expect(() => service.getCpuProfile(null, null), throwsUnimplementedError);
     });
 
     test('getFlagList', () {
@@ -620,6 +608,16 @@ void main() {
       await resumeSub.cancel();
     });
 
+    test('getInboundReferences', () async {
+      expect(() => service.getInboundReferences(null, null, null),
+          throwsUnimplementedError);
+    });
+
+    test('getRetainingPath', () async {
+      expect(() => service.getRetainingPath(null, null, null),
+          throwsUnimplementedError);
+    });
+
     test('registerService', () async {
       expect(() => service.registerService('ext.foo.bar', null),
           throwsUnimplementedError);
@@ -627,11 +625,6 @@ void main() {
 
     test('reloadSources', () {
       expect(() => service.reloadSources(null), throwsUnimplementedError);
-    });
-
-    test('requestHeapSnapshot', () {
-      expect(() => service.requestHeapSnapshot(null, null, null),
-          throwsUnimplementedError);
     });
 
     test('setExceptionPauseMode', () async {