+++ dwds/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix an issue where we source map paths were not normalized.
 - Added a check to tests for the variable DWDS_DEBUG_CHROME to run Chrome with a
   UI rather than headless.
+- Depend on latest `package:vm_service`.
 
 ## 0.4.0